Changed site_url() to home_url()
*site_url():*
Retrieves the URL for the current site where WordPress application files (e.g. wp-blog-header.php or the wp-admin/ folder) are accessible.

*home_url():*
Retrieves the URL for the current site where the front end is accessible.

If you have installed WordPress in its own directory the site_url() includes it in the URL. This results in a 404.
I changed this to the home_url. This should work properly regardless of the WordPress directory.